Add extensive unit tests and bug fixes
Add many JUnit tests across the codebase and fix several defects. Highlights:
- Add maven-surefire-plugin with --enable-preview and include /.maven in .gitignore.
- Pow: implement equalsTo; add PowTest.
- Root: handle negative arguments in eval; add RootTest.
- Conversions: reverse binary digit list before joining.
- Point3D: compare coordinates using doubleValue() to avoid boxing issues.
- Vector2D/Vector3D: remove incorrect (long) casts when computing angles; Vector3D now implements Cloneable and fixes sum/subtract to use destination coordinates correctly.
- Numerous new/updated tests for algebra, exponentials, logarithmics, operators, vectors, points and other utilities to increase coverage and catch regressions.

+++ src/main/java/MathUtils/point/Point3D.java
@@ -27,9 +27,9 @@ public final class Point3D extends Point<Point3D, Vector3D> {
 
     @Override
     public boolean isEqualTo(Point3D point) {
-        return point.x == getX() &&
-                point.y == getY() &&
-                point.z == getZ();
+        return point.x.doubleValue() == getX().doubleValue() &&
+                point.y.doubleValue() == getY().doubleValue() &&
+                point.z.doubleValue() == getZ().doubleValue();
     }
 
     public Double getX() {
